+++ CHANGELOG.md
@@ -14,7 +14,12 @@ Types of changes are:
 
 ## Unreleased
 
+### Added
+- New system icons
+- Add the new system icons to `IconFactory Component`
+
 ## 2.0.12 - 2025-09-22
+
 ### Added
 - New system icon (download)